refactor(assistant): move app-store out of memory/ into apps/
