+++ assistant/src/__tests__/app-routes-csp.test.ts
@@ -34,7 +34,7 @@ const apps = new Map<string, typeof legacyApp | typeof multifileApp>([
   ["multi-1", multifileApp],
 ]);
 
-mock.module("../memory/app-store.js", () => ({
+mock.module("../apps/app-store.js", () => ({
   getApp: (id: string) => apps.get(id) ?? null,
   getAppsDir: () => "/fake/apps",
   getAppDirPath: (appId: string) => `/fake/apps/${appId}`,
