+++ assistant/src/__tests__/conversation-runtime-workspace.test.ts
@@ -1,12 +1,12 @@
 import { afterEach, describe, expect, test } from "bun:test";
 
+import { createApp } from "../apps/app-store.js";
 import {
   clearConversations,
   setConversation,
 } from "../daemon/conversation-registry.js";
 import { applyRuntimeInjections } from "../daemon/conversation-runtime-assembly.js";
 import type { SurfaceData, SurfaceType } from "../daemon/message-protocol.js";
-import { createApp } from "../memory/app-store.js";
 import { registerDefaultPluginInjectors } from "../plugins/defaults/index.js";
 import type { Message } from "../providers/types.js";
