+++ assistant/src/__tests__/conversation-workspace-tool-tracking.test.ts
@@ -201,7 +201,7 @@ mock.module("../plugins/defaults/compaction/window-manager.js", () => ({
 mock.module("../persistence/llm-usage-store.js", () => ({
   recordUsageEvent: () => ({ id: "usage-1", createdAt: Date.now() }),
 }));
-mock.module("../memory/app-store.js", () => ({
+mock.module("../apps/app-store.js", () => ({
   getApp: () => null,
   updateApp: () => {},
 }));
