+++ assistant/src/__tests__/processing-flag-persist-failure.test.ts
@@ -185,7 +185,7 @@ mock.module("../persistence/llm-usage-store.js", () => ({
   recordUsageEvent: () => ({ id: "usage-1", createdAt: Date.now() }),
 }));
 
-mock.module("../memory/app-store.js", () => ({
+mock.module("../apps/app-store.js", () => ({
   getApp: () => null,
   updateApp: () => {},
 }));
