+++ assistant/src/config/bundled-skills/app-builder/tools/app-refresh.ts
@@ -1,5 +1,5 @@
 import { setAppCommitMessage } from "../../../../apps/app-git-service.js";
-import * as appStore from "../../../../memory/app-store.js";
+import * as appStore from "../../../../apps/app-store.js";
 import { executeAppRefresh } from "../../../../tools/apps/executors.js";
 import {
   missingAppIdError,
