+++ assistant/src/config/bundled-skills/app-builder/tools/app-update.ts
@@ -1,5 +1,5 @@
 import { setAppCommitMessage } from "../../../../apps/app-git-service.js";
-import * as appStore from "../../../../memory/app-store.js";
+import * as appStore from "../../../../apps/app-store.js";
 import type { AppUpdateInput } from "../../../../tools/apps/executors.js";
 import { executeAppUpdate } from "../../../../tools/apps/executors.js";
 import {
