+++ assistant/src/daemon/conversation-surfaces.ts
@@ -7,7 +7,6 @@ import {
   CardSurfaceDataSchema,
   FileUploadSurfaceDataSchema,
 } from "../api/surfaces.js";
-import { isActivationSession } from "../memory/activation-session-store.js";
 import {
   addAppConversationId,
   getApp,
@@ -18,7 +17,8 @@ import {
   resolveAppDir,
   resolveEffectiveAppHtml,
   updateApp,
-} from "../memory/app-store.js";
+} from "../apps/app-store.js";
+import { isActivationSession } from "../memory/activation-session-store.js";
 import { recordActivationEvent } from "../onboarding/onboarding-events-store.js";
 import {
   getMessages,
