+++ assistant/src/daemon/tool-side-effects.ts
@@ -9,8 +9,8 @@
 
 import { isAbsolute, resolve, sep } from "node:path";
 
+import { addAppConversationId } from "../apps/app-store.js";
 import { generateAppIcon } from "../media/app-icon-generator.js";
-import { addAppConversationId } from "../memory/app-store.js";
 import { invalidateEdgeIndex } from "../memory/v2/edge-index.js";
 import { invalidatePageIndex } from "../memory/v2/page-index.js";
 import { getConceptsDir } from "../memory/v2/page-store.js";
