+++ assistant/src/runtime/routes/publish-routes.ts
@@ -14,17 +14,17 @@ import { v4 as uuid } from "uuid";
 import { z } from "zod";
 
 import {
+  getApp,
+  getAppDirPath,
+  isMultifileApp,
+  resolveEffectiveAppHtml,
+} from "../../apps/app-store.js";
+import {
   getActivePublishedPageByAppId,
   insertPublishedPage,
   updatePublishedPage,
 } from "../../apps/published-pages-store.js";
 import { compileApp } from "../../bundler/app-compiler.js";
-import {
-  getApp,
-  getAppDirPath,
-  isMultifileApp,
-  resolveEffectiveAppHtml,
-} from "../../memory/app-store.js";
 import { deployHtmlToVercel } from "../../services/vercel-deploy.js";
 import { credentialBroker } from "../../tools/credentials/broker.js";
 import { getLogger } from "../../util/logger.js";
